Move hacky code in specific subclass (#46)
