+++ lib/src/main/java/com/tbruyelle/rxpermissions/ShadowActivity.java
@@ -8,26 +8,12 @@ import android.os.Bundle;
 import android.os.Process;
 
 @TargetApi(Build.VERSION_CODES.M)
-public class ShadowActivity extends Activity {
-    private static final String KEY_ORIGINAL_PID = "key_original_pid";
-
-    private int mOriginalProcessId;
-
+public class ShadowActivity extends EnsureSameProcessActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         if (savedInstanceState == null) {
             handleIntent(getIntent());
-
-            mOriginalProcessId = Process.myPid();
-        } else {
-            mOriginalProcessId = savedInstanceState.getInt(KEY_ORIGINAL_PID, mOriginalProcessId);
-
-            boolean restoredInAnotherProcess = mOriginalProcessId != Process.myPid();
-
-            if(restoredInAnotherProcess) {
-                finish();
-            }
         }
     }
 
@@ -45,12 +31,5 @@ public class ShadowActivity extends Activity {
     public void onRequestPermissionsResult(int requestCode, String[] permissions, int[] grantResults) {
         RxPermissions.getInstance(this).onRequestPermissionsResult(requestCode, permissions, grantResults);
         finish();
-    }
-
-    @Override
-    protected void onSaveInstanceState(Bundle outState) {
-        super.onSaveInstanceState(outState);
-
-        outState.putInt(KEY_ORIGINAL_PID, mOriginalProcessId);
     }
 }
